fix: Dont use CDKTF_ prefix on env variables that are not for all commands

+++ test/local-registry.sh
@@ -35,5 +35,7 @@ function stopLocalRegistry {
   # Restore the original NPM and Yarn registry URLs and stop Verdaccio
   npm set registry "$original_npm_registry_url"
   yarn config set registry "$original_yarn_registry_url"
-  kill $verdaccio_pid
+  listener_pid=`lsof -i :4873 -t`
+  [ -n "$listener_pid" ] && kill $listener_pid
+  [ -n "$verdaccio_pid" ] && kill $verdaccio_pid
 }
